# Backend/github-discovery/output/20edbcb95ceffb10_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Specification - 20edbcb95ceffb10",
    "version": "1.0.0",
    "description": "Auto-generated from repository https://github.com/ilyeschaabani/microservice.git",
    "x-generated": "2026-03-23T16:12:12.162345Z",
    "x-endpoint-count": 34
  },
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ],
  "paths": {
    "/api/auth/api/auth/signup": {
      "post": {
        "summary": "signUp",
        "tags": [
          "sign",
          "authentification_rest_a_p_i"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/signup\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/auth/api/auth/signin": {
      "post": {
        "summary": "signIn",
        "tags": [
          "sign",
          "authentification_rest_a_p_i"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/signin\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/auth/api/auth/refreshToken": {
      "post": {
        "summary": "refreshToken",
        "tags": [
          "refresh",
          "authentification_rest_a_p_i"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/refreshToken\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/forgotPassword/forgotPassword/verifyMail/{email}": {
      "post": {
        "summary": "verifyEmail",
        "tags": [
          "verify",
          "forgot_password"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/verifyMail/{email}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/forgotPassword/forgotPassword/verifyOTP/{otp}/{email}": {
      "post": {
        "summary": "verifyOTP",
        "tags": [
          "verify",
          "forgot_password"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/verifyOTP/{otp}/{email}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/forgotPassword/forgotPassword/changePassword/{email}": {
      "post": {
        "summary": "changePasswordHandler",
        "tags": [
          "change",
          "forgot_password"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/changePassword/{email}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/formations/api/formations/add": {
      "post": {
        "summary": "addFormation",
        "tags": [
          "add"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/add\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/formations/api/formations/update/{id}": {
      "put": {
        "summary": "updateFormation",
        "tags": [
          "update"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PutMapping(\"/update/{id}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/formations/api/formations/delete/{id}": {
      "delete": {
        "summary": "deleteFormation",
        "tags": [
          "delete"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@DeleteMapping(\"/delete/{id}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/formations/api/formations/all": {
      "get": {
        "summary": "getAllFormations",
        "tags": [
          "get"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@GetMapping(\"/all\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/formations/api/formations/get/{id}": {
      "get": {
        "summary": "getFormationById",
        "tags": [
          "get"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@GetMapping(\"/get/{id}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}